examples/python/requirements.txt
lightrace
python-dotenv>=1.0
anthropic>=0.39
openai>=1.0
langchain-core>=0.2
langchain-openai>=0.2
